change db from mysql to postgresql

+++ src/main/java/com/demo/jersey/service/UserApi.java
@@ -15,7 +15,7 @@ public class UserApi {
 	@GET
 	@Path("/{status}/count")
 	@Produces(MediaType.TEXT_PLAIN)
-	public int getUserCountByStatus(@PathParam("status") final Short status) {
+	public int getUserCountByStatus(@PathParam("status") final int status) {
 		return userRepository.getUserCountByStatus(status);
 	}
 }
